Utils file
Also switched to using a pure function,
`copyImage(svg, color).then(itWorked, itFailed)`
